agri_call_agent/requirements.txt
fastapi
uvicorn
requests
python-dotenv
jinja2
edge-tts